pyproject.toml
[tool.poetry]
name = "service_download"
version = "2.2.0"
description = ""
authors = ["Indoc Research"]

[tool.poetry.dependencies]
aiofiles = "^0.8.0"
aiokafka = "^0.7.2"
aioredis = "^2.0.1"
anyio = "3.4.0"
asyncpg = "0.25.0"
fastapi = "^0.79.0"
fastapi-health = "^0.4.0"
fastapi-utils = "0.2.1"
fastavro = "^1.5.2"
greenlet = "^1.1.2"
gunicorn = "20.0.4"
uvloop = "0.17.0"
httptools = "0.5.0"
python-multipart = "0.0.5"
pydantic = "1.8.2"
opentelemetry-exporter-jaeger = "1.6.2"
opentelemetry-instrumentation = "0.27b0"
opentelemetry-instrumentation-asyncpg = "0.27b0"
opentelemetry-instrumentation-fastapi = "0.27b0"
opentelemetry-instrumentation-httpx = "0.27b0"
opentelemetry-instrumentation-requests = "0.27b0"
opentelemetry-instrumentation-sqlalchemy = "0.27b0"
pilot-platform-common = "^0.1.3"
pyjwt = "^2.6.0"
psycopg2-binary = "2.9.2"
python = "^3.7"
sqlalchemy = {extras = ["asyncio"], version = "^1.4.32"}
uvicorn = "0.12.3"

[tool.poetry.dev-dependencies]
async-asgi-testclient = "^1.4.9"
Faker = "11.3.0"
flake8 = "^4.0.1"
pytest = "7.1.2"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.18.3"
pytest-mock = "^3.7.0"
pytest-httpx = "^0.21.0"
testcontainers = "^3.4.2"

[tool.pytest.ini_options]
testpaths = "tests"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
